youtube_uploader/requirements.txt
google-auth-oauthlib==1.2.0
google-auth==2.25.2
google-api-python-client==1.12.5
python-dotenv==1.0.0
requests==2.31.0
